Move type back into Grid
