+++ packages/react/src/Grid/Grid.tsx
@@ -10,6 +10,7 @@ import { GridCell } from './GridCell'
 
 export type GridColumnNumber = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12
 export type GridPaddingSize = 'small' | 'medium' | 'large'
+export type ResponsiveGridValues<T> = T | { narrow: T; medium: T; wide: T }
 
 type GridPaddingVerticalProp = {
   paddingBottom?: never
