--- packages/react/src/common/types.ts
@@ -1,1 +1,0 @@
-export type ResponsiveGridValues<T> = T | { narrow: T; medium: T; wide: T }
